style(global): satisfy dylint (external test files, raw strings, single-letter renames)
