+++ pacquet/crates/cli/src/cli_args/global.rs
@@ -46,12 +46,12 @@ pub enum GlobalError {
     #[diagnostic(
         code(ERR_PNPM_NO_GLOBAL_BIN_DIR),
         help(
-            "Run \"pnpm setup\" to create it automatically, or set the global-bin-dir setting, or the PNPM_HOME env variable. The global bin directory should be in the PATH."
+            r#"Run "pnpm setup" to create it automatically, or set the global-bin-dir setting, or the PNPM_HOME env variable. The global bin directory should be in the PATH."#
         )
     )]
     NoGlobalBinDir,
 
-    #[display("Use the \"pnpm self-update\" command to install or update pnpm")]
+    #[display(r#"Use the "pnpm self-update" command to install or update pnpm"#)]
     #[diagnostic(code(ERR_PNPM_GLOBAL_PNPM_INSTALL))]
     GlobalPnpmInstall,
 
@@ -442,30 +442,4 @@ fn is_windows_drive_path(param: &str) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{is_windows_drive_path, split_comma_separated};
-    use std::path::Path;
-
-    #[test]
-    fn comma_splits_into_selectors() {
-        let base = Path::new("/nonexistent");
-        assert_eq!(split_comma_separated("foo,bar", base), vec!["foo", "bar"]);
-        assert_eq!(split_comma_separated("foo", base), vec!["foo"]);
-    }
-
-    #[test]
-    fn urls_are_kept_whole() {
-        let base = Path::new("/nonexistent");
-        assert_eq!(
-            split_comma_separated("https://example.com/a,b.tgz", base),
-            vec!["https://example.com/a,b.tgz"]
-        );
-    }
-
-    #[test]
-    fn detects_windows_drive_paths() {
-        assert!(is_windows_drive_path("C:\\foo"));
-        assert!(is_windows_drive_path("d:/bar"));
-        assert!(!is_windows_drive_path("foo"));
-    }
-}
+mod tests;
