+++ pacquet/crates/config/src/lib.rs
@@ -8,8 +8,10 @@ mod store_path;
 pub mod version_policy;
 mod workspace_yaml;
 
-pub use crate::api::{EnvVar, EnvVarOs, GetCurrentDir, GetHomeDir, Host, LinkProbe};
-pub use crate::global_bin_check::{CheckGlobalBinDirError, check_global_bin_dir};
+pub use crate::{
+    api::{EnvVar, EnvVarOs, GetCurrentDir, GetHomeDir, Host, LinkProbe},
+    global_bin_check::{CheckGlobalBinDirError, check_global_bin_dir},
+};
 
 use indexmap::IndexMap;
 use pacquet_patching::{
