+++ pacquet/crates/global/src/list.rs
@@ -308,16 +308,4 @@ fn gray(text: &str) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::glob_match;
-
-    #[test]
-    fn glob_matches_exact_and_wildcards() {
-        assert!(glob_match("foo", "foo"));
-        assert!(!glob_match("foo", "bar"));
-        assert!(glob_match("*", "anything"));
-        assert!(glob_match("@scope/*", "@scope/pkg"));
-        assert!(!glob_match("@scope/*", "@other/pkg"));
-        assert!(glob_match("foo*bar", "fooXYZbar"));
-    }
-}
+mod tests;
